+++ lib/src/references/references.dart
@@ -1,4 +1,4 @@
-// Copyright (C) 2024-2025 Intel Corporation
+// Copyright (C) 2024-2026 Intel Corporation
 // SPDX-License-Identifier: BSD-3-Clause
 
 import 'package:collection/collection.dart';
@@ -8,6 +8,7 @@ import 'package:rohd_bridge/rohd_bridge.dart';
 import 'package:rohd_bridge/src/references/reference.dart';
 
 part 'port_reference.dart';
+part 'typed_port_reference.dart';
 part 'interface_reference.dart';
 part 'slice_port_reference.dart';
 part 'standard_port_reference.dart';